# Locale performance improvements