--- a/components/locid/src/extensions/transform/value.rs
+++ b/components/locid/src/extensions/transform/value.rs
@@ -123,7 +123,10 @@ fn test_writeable() {
     let foobar = "foobar".parse().unwrap();
 
     assert_writeable_eq!(Value::default(), "true");
-    assert_writeable_eq!(Value::from_short_slice_unchecked(vec![hybrid].into()), "hybrid");
+    assert_writeable_eq!(
+        Value::from_short_slice_unchecked(vec![hybrid].into()),
+        "hybrid"
+    );
     assert_writeable_eq!(
         Value::from_short_slice_unchecked(vec![hybrid, foobar].into()),
         "hybrid-foobar"