Specifically improving `ShortSlice::retain` and `ShortSlice::from_iterator` to be less allocatey for single-element slices.

Replacing all `Vec` usages in `icu_locid` by `ShortSlice`. This will be less performant for constructing long lists, but more performant for length-1 lists, as those can be constructed allocation-free.

```
langid/overview         time:   [1.5500 µs 1.5515 µs 1.5531 µs]
                        change: [-2.2337% -1.9994% -1.7885%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

     Running benches/locale.rs (target/release/deps/locale-a81f3922e113b14b)
locale/overview         time:   [2.7486 µs 2.7517 µs 2.7554 µs]
                        change: [-5.9166% -5.7892% -5.6582%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
```